chore: bump deps to avoid event-stream & pin hugo
nodemon was potentially affected by the event-stream issue
https://github.com/dominictarr/event-stream/issues/116

This PR bumps those deps to the latest release that avoid it.

To make the site build reliably, it adds hugo as a dev dep and
pins it to vesion v0.31.1 that is known to work with the current
ipld.io website src.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>