Expand more of the Type member types.
Added some enums that should be useful, and documented the bejezus out
of them.

Seems to be going well so far.

Documention for TypeUnion.

All the other interesting Type types are still todos, though; and we
thus far haven't hoisted ourselves back up to the point where typeterm
recursiveness needs to be fleshed out, which is one of the things we
already know is going to be Fun.

Fixed what's probably a typo with an extra equals sign.  I'm not sure
why my brain keeps wanting to do that for unions.

And one interesting note: though I'm keeping TypeKind in the commit,
see the todo comment about its potential irrelevance?  It's amusing
that we've just parallelly reproduced that; the same misstep already
happened once before, just in the golang code: TypeKind there was also
not needed, because the Type union members already represented it.

Signed-off-by: Eric Myhre <hash@exultant.us>