Define TypeEnum and expand json schema using it.
Not having a "set" definition is kind of irritating.

I'm wondering if we should propose to make a set kind; and simply
transform these to a map with null values as necessary.  Possibly it
could even be argued that this could make sense at the Data Model
level?  I'm not sure.  Something that should be discussed, though.

Signed-off-by: Eric Myhre <hash@exultant.us>