A proposal for TypeTerm, nullable, InlineDefn, etc
This seems to work and capture all of the concepts that we want:

- the inline definitions are possible;
- they recurse correctly
- nullability can be inserted in all the correct places;
- you *can't* have a "nullable nullable" nesting;
- and bonus, the simple TypeName case is still a string repr.

The downsides are:

- wow, this *looks* complicated on the schema side;
- in part because we have a union-union nest, and conceptually we'd
   probably simply prefer to unify those, but it's a representational
   problem (can't clearly combine a keyed and a kinded);
- the json repr looks to get deep fast.

I wonder if we can do better than this.

And also, we shifted the flagging of optional fields to
StructRepresentation_Map, then eliding a struct for field properties
entirely.  This seemed potentially sensible given that we suppose a
struct can be represented as a list, in which case optionality of
fields won't work the same.  However, on further thought, I'm also
extremely dubious of this idea: whether or not a field can be traversed
and then yield "undefined" is defined by whether it's optional; and
I'm really suspecting it's going to create a lot of weirdness to make
that feature conditional upon the StructRepresentation union.
If anything, we should simply make invalid combinations of optionality
and representation of that struct, well, erm, invalid.

Signed-off-by: Eric Myhre <hash@exultant.us>