Another, flatter proposal for TypeTerm, etc.
(An alternative proposal, with the same example DSL but different
choices in schema representation is in a sibling commit, and we'll
merge and pick one or the other to continue with shortly.)

This take on TypeTerm keeps things flatter and has fewer schema types:
the optional and nullable properties are those of the containers
in the recursive type (i.e., the map itself, list itself, or the fields
in the case of structs).

As with the alternative proposal, this seems to work and capture all
of the concepts that we want:

- the inline definitions are possible;
- they recurse correctly
- nullability can be inserted in all the correct places;
- you *can't* have a "nullable nullable" nesting;
- and bonus, the simple TypeName case is still a string repr.

So, since all the same key properties exist in both approaches,
in some ways, this is six-of-one/half-a-dozen-of-the-other.
Both mechanisms of encoding clearly represent the same things.
It's just question of which is nicer to work with, which seems more
self-documenting to a human reader, and other intangibles of that sort.

In this approach, we have a less rightward-leaning tree when the schema
is represented in JSON form; by comparison, the alternative which had
"nullable" represented as a union experiences deeper nested objects.
Reducing depth-of-tree is hardly a proof of simplicity, but it's usually
a good heuristic, so we can count that as a win here.

Another comparison we can make is that the number of types used in this
diff is 2; the number used in the other approach was 4.  We do this
at the trade of more fields in existing types, of course, so it's less
than perfectly clear cut as a win for simplicity... but I tend towards
thinking it is.  Especially certainly clearer is that we *don't* have
a union type which immediately nests another different representation
of union type in this approach!

So, while it's always hard to make absolutionist statements about
"simplicity", indeed, this seems simpler via several observations.

In terms of what other people are doing...
One interesting example from another endevour I recently noticed -- the
Atom project emits a JSON file describing their API:
https://github.com/atom/atom/releases/download/v1.34.0/atom-api.json
In this JSON file, they've taken a tack similar to the one emboded in
this commit: the "isOptional" property is inline with the field defn's.
So, we can count Atom's choice as an effective a +1 vote for that.
(It's perhaps interesting they also have *names* in their field obj...
however, the fields here are in an array, which is necessary since
they're describing positional arguments.  So, that's entirely
self-consistent, but also different constraints than we have, and thus
while interesting, not quite comparable to our choices on that matter.)

Note that the JSON-formatted schema-schema file has not been updated yet.
We'll update and finish that after merging and selecting a proposal here.

Signed-off-by: Eric Myhre <hash@exultant.us>