More optional properties (for optionals!).
The JSON representation of the schema shortens up noticeably if the
properties for "nullable" and "optional" are themselves optional.

Given that the hope is for the in-practice use of these properties to
be the exceptional case rather than the norm, it's nicer to speak
less of them.

This edges around an interesting question: we don't actually *want*
three-valued logic here... but that's precisely what "optional Bool"
gives you: it may have the value "true", "false", or "undefined"!
What we *mean* here is of course "true" or "undefined".  It would be
desirable to have a way to encode that.  Perhaps we can find a way.
One idea is that this is an issue in the representation of bool;
and thus perhaps we should configure it in the struct's representation
block rather than marking it optional in the struct's definition block.
Let's give this more thought in the future.

Signed-off-by: Eric Myhre <hash@exultant.us>