Data Model doc expanded, and in terms of Kinds.
This is part of a motion towards being able to introduce additional
new terminology as a part of work on Schemas (which will follow in
(many) additional separate commits) -- which is alluded to in the
content already.

(There's also a dangling link to a 'schema' folder that doesn't exist
yet, but will quite presently.  Mea cupla; gotta pick *some* order to
get started in!)

Signed-off-by: Eric Myhre <hash@exultant.us>