Big delicious union representation examples.
Union representations are a fun topic.  It's something with very
minimal prior art that I'm aware of -- systems with *unions* exist,
surely; and systems which have informal but practical techniques
which are *regarded* as unions exist; but systems which attempt to both
formalize the concept AND do so in terms of existing formats without
introducing a new distinct serialization system are rare indeed.

Signed-off-by: Eric Myhre <hash@exultant.us>