Schema update: use implicits; also, more docs.
- the "default" feature of representations has been renamed:
it is now known as "implicit".  We believe this better clarifies
what the feature does: the old name suggests that it might create
a surjective function; it is in fact a bijective function.
(Thanks to @merovius for the first suggestion of this new name!)

- several representation blocks removed: we've introduced the
ability to annotate fields in struct types with per-field
representation details inline, which shortens some things
significantly.  (The IPLD-native representation still keeps
representations clearly separated even at the cost of redundant keys.)

- new documentation file describing the above design.