Schema update: no more anon params to repr strat.
Previous syntax which allowed anonymous parameters to representation
strategies dropped.  (This had appeared in the schema-schema in the
uses of "inline" unions.)  Those parameters are now always named, and
thus require the use of a full block to complete the representation
description.  This is much more future-proof, and gives more
indications about what is going on, as well as more words we can
ascribe documentation to.

This change, like most of the others I'm making today, is prompted by
multiple incidents of user feedback indicating that the earlier syntax
was non-obvious (demonstrably: it forced them to ask questions).