Schema update: remove obvious/redundant kind decl.
Elide 'map' and 'list' keywords in top-level type declarations.

In all other positions (e.g. inline type declarations for struct
fields, nested maps, and nested lists), there's no kind keyword used:
the '{' and '[' characters in the typeterm are already enough to be
explicit about what's going on.  Historically, we've also added
a kind keyword when making top-level type declarations, even though
it's technically redundant.

I'm a bit tepid about this change.  However, many people have
commented that the 'map' and 'list' keywords in this position seem odd
and redundant.  I'm outvoted.

The original idea was that the form is always roughly
"type $typeName $kind [$block...]" -- and the "block" would be
present only in types that need more info (like structs).
But it's not much of a pivot to just switch the model to
"type $typeName $typeDefn" -- where there's not necessarily an
appearance of a kind keyword, and instead when we see words like
"struct" appear here, it's because it's part of the "typeDefn".

So.  Here's that take.