Exploration notes from thinking about int ranges.
Numeric precision is a tricky issue.

As with other files added to this directory, this is an "exploration report"
and it contains a series of "in the opinion of this author" sorts of comments.

I'm not sure this one came to a great conclusion, but it contains at least
one interesting probably-good idea, and a variety of useful references for
future iterations.

Since I can't easily proceed much further with this document without mentally
"crumpling it up" and starting a fresh one, it seems like at this point I
ought to checkpoint and share, to see if anyone else has ideas.
If there is a followup document to this one, it might refer to or crib
snippets from this one, but will overall need a clean start.