Factor out Selector design history from present.
There's a lot of spectacular content here, but recent discussion
( https://github.com/ipld/specs/pull/133#discussion_r302860028 ,
https://github.com/ipld/specs/pull/141#issuecomment-510777921 ,
and other various out-of-band conversation) has highlighted the fact
that the main file describing Selectors has grown a bit unwieldy,
and contains a confusing mixture of present-tense fact as well as
historical (and now often resolved!) questions and design thought.

Therefore, this commit attempts to extricate the two.

All of the historical stuff goes to a file in the  `./design-history`
folder, and I've written a new suffix for it which links to a
variety of the other in-progress stages things have passed through.

The main selectors doc is now if anything a bit *too* skeletal -- we
need to now write succinct, modern introductions and descriptions!
(I'll continue this in subsequent commits -- as well as adding back
new up-to-date examples and fixtures.)