Tweak foundations definitions: path segments, etc.
Dropped mention of "Fragments".  We don't seem to use that term much.
It's not a bad term; it just doesn't seem to have reached the frequency
of usage which makes it seem important to include in the glossary of
most critical terms.

Inserted the word "kind" into the description of links.

Reworked "Path" to describe "Path" and "Path Segment" separately.
This bit is a bit touchy to phrase well, and has already gone through
some massaging in another PR; hopefully we're getting *closer*.