Readme covering intent of design history dir.
It seems like covering what is and is not imagined to go under this
umbrella is going to be a FAQ, so... I wrote an FAQ.

Some, though not all, of this harkens back to
https://gist.github.com/warpfork/8c50277bc48fbd0591bae326099a9c18 .
(So, in a way you can surely see *I'm serious* about gists being a
good place to incubate things!)

The concept of a "lab notebook" also gets serious homage here.
Probably worth writing about that more in its own right.
But this will do, here, for now.