Introductory document to how IPLD fits with other Interplanetary projects.
I also made a "meta" file which briefly remarks on what I wanted to
accomplish with this document.  Not sure if this will be useful in the
long run, but it's an experiment I want to try.  We seem to have a lot
of discussions about what document purpose and scope is when editing;
maybe talking about that explicitly, and in version control (so we
don't forget about it when revisiting documents over time!), will be a
useful idea.  At the same time, doing so inline in the same file
would create a lot of line noise for other readers.  Therefore: a
separate (but adjacent) file seems like it may be suitable.