Vision and scope documents.
I am frankly riffing these off the top of my head.  These are the
most important things I could think of, and the things that connect
most closely to our actual roadmap and the reasoning behind it.
I may have forgotten some things, though.  And I've no doubt we'll
want to rehash the phrasing.  I wax verbose.

Still, it's a start on content we need.  Here it is.