Attempt to describe component dependencies in dot.
Everything I'm making a node for here is something I expect to
dedicate a full page to in this folder.

Imperfect, first draft, not pretty; we'll see if it sticks.

I'm thinking it may be useful in the future to generate this as an SVG
and turn each node into a link, but we'll cross that bridge when/if we
get to it.

The command to regenerate this png is familiar to any user of graphviz:

`dot component-dependencies.dot -T png > component-dependencies.png`