sift apart beginner tutorials from user stories.
might not be the final dir or link names; still kind of fumbling for
the right level of description for what content like this is.

dropped "files in ipld" stub file because while that's surely something
we want to describe somewhere, that stub isn't helping and i think it
probably won't go in this section at all (at the moment, I'm thinking
I might have some content ready for the "ecosystem" dir tree).