block-layer/codecs: dag-pb can't enforce PBNode ordering
In general, Protobuf encoders will encode message fields in increasing
order of field numbers. This would mean Data before Links.

Unfortunately, IPFS's PBNode encoder in go-merkledag started being used
with a bug: it would encode Links before Data.

Since changing that ordering would result in hash and CID mismatches,
the encoding order was kept when marshaling. The same order is reflected
in how dag-pb has always required the data be encoded.

The second unfortunate event is that some end users started using their
own PBNode encoders, with protobuf generators which would follow the
"field number" order, opposed to IPFS's own encode order.

Since most protobuf decoders accept any order - and this includes
go-merkledag's - it means that both PBNode orders are already relatively
widespread in the public set of IPFS data.

As such, we must amend the spec to remain backwards compatible and be
used in go-ipfs.

The existing rule is changed to refer to PBLink, as it has no such
ambiguity, and we are not aware of any alternative ordering out in the
wild. Plus, the order defined in the spec matches Protobuf's default
order, too.

A new rule is added for PBNode. It clarifies that the encode order
"must" be different than the field number order, and specifies that the
decode "should" accept either order.

The wording for the new rule is a mix of "must" and "should", since we
want to strongly encourage software to use a canonical form when
encoding for the sake of CIDs, but we must keep backwards compatibility
when decoding old data.