site tooling and some early content outlines.
I'm going with eleventy as a static site generator after a great deal
of deliberation and consideration of numerous other tools.
It's very direct, it's very fast, it's reasonably extensible,
it's pleasingly non-magical, and it does the key things without a fuss.

As of this commit, the output is not well decorated :)  CSS will come.

I'm using a makefile as the entry point because I'm unwilling to have
this repo look like it's a javascript project, and expect to integrate
several other kinds of tooling in various langauges over time
(starting, already, with a link checker tool).

The configuration files for eleventy are all located in the ".site"
directory (and this does include npm package files, inevitably).
This seems to be a mildly nonstandard arrangement, but demonstrably
works, and has a nice clean result.
(The "eleventyignore" file is the one exception.  I haven't figured
out an incantation for this which does the right things if located
elsewhere.  Oh well.  This is livable.)

I'm using https://github.com/untitaker/hyperlink for link checking.
Again, I've weighed several alternatives.  This one is the fastest,
and the most feature complete (in particular, it can check in-page
anchor targets, which I consider a key feature, given how heavily we
use them), and it works very cleanly on *files* alone, meaning it
will be safe and easy to use in a CI context (there's no dialing out to
check other web resources, which would be both a source of wasted time
and a source of flakiness not actually related to diffs).

A publication/deploy script is not yet included.  Future work.
Meanwhile, the output is in ".site/_output/".