the essentials of a theme appear!
I'm doing this from scratch in an extremely minimalist fashion,
and lots of comments (particularly about the margin magic),
and trying to avoid javascript as much as possible.

(One piece of javascript has crept in; I hope it's removable in the
future, but it's a bit more of a quest than I care to chase today.)

I'm riffing off of the theme and offsets used in the old site,
but not any of its source code directly... because frankly,
I can't find it.

Haven't actually filled out any of the sidebar nav of suchlike yet.
Soon(tm).