Begin use of eleventy navigation plugin to help generate a nav menu.
This is... a bit of a journey.

First of all, it's caused me to switch to njk templates instead of liquid.
In large part, this is because liquid doesn't seem to want to have recursion features,
and that makes a nonstarter for handling things like a menu that can have variable depth.
(See https://github.com/Shopify/liquid/issues/580 ; from this, I gather that it's
not just that liquid doesn't have a feature for this, it's that they don't *want* one.
I understand the desire to avoid recursion (and thus turing-completeness), but it's unhelpful here.)
In smaller part, it's just going with the flow of eleventy docs, which seem to prefer njk.
Ultimately, the strength of my preference in *any* direction would be weak,
but liquid crossed a hard-disqualifying line, so, let's try njk now.

Okay.  Moving on.  All that was just to describe why one file is renamed.

About this nav plugin:

- It works!

- I like having fairly explicit control over what ends up in nav,
  and how its labeled.  (I expect we'll have quite a few pages, so,
  being able to be explicit about what's listed and how will be
  import to be able to control the signal-noise of the display.)

- Ordering in the list is controlled by numbers in the per-page frontmatter.
  I'm not really a fan of this at all.  I'd rather nudge around an overall list.
  Oh well; this isn't critical enough to spend a lot of time on.

- I'm super unthrilled by having to repeat the parent key a lot.
  I think this navigation plugin is trying to support the ability to
  have navigational hierarchy that's totally divorced from directory structure...
  but _I don't really want that_, and it makes basic use more redundant.

- I feel like I might be able to improve the above by computed properties
  based on the page URL.  But I haven't discerned an incantation for that yet.

- One can also use the `foo/foo.11tydata.json` file to set the nav parent
  property for the whole directory... and maybe I'll take that angle...
  but I'd still find this relatively high friction and a bit disappointing.

- I'm a tad worried this nav thing is gonna get me in trouble when it collides
  with the intention to sometimes have the same page names in parallel in
  the docs and the specs directory trees (e.g. for those things that need
  both a formal spec as well as a human-friendly explainer doc).
  The likely result will be I'll stop using human-readable keys for the
  nav config, and use separate title properties...?  Mmfh.

- I think I'm going to be using CSS on the nav menu *heavily* to control
  the legibility and noise level of things.  Probably including setting
  all elements beyond a certain depth to display-none.  (And then adding
  back other css selectors that can toggle them back into visibility.)
  Future work.

So it feels like I'm going to be reinventing a good bit of kludge to
make this navigation plugin work in the simple way that I want.
But maybe that's okay.  Let's see how it goes.

Worst case scenario?  I might make a totally static nav menu, by hand
(with just a little template power sprinkled on for the you-are-here
highlighting based on the current page url).
I'd like to have data that can be used to generate group listings, too
(for example, as you can see being formed for the "known codecs" page),
but maybe that can be a separate problem anyway (using "categories"
instead of a nav tree for that might be appropriate)).

Anyway.  Let's see how it goes.