Successfully automated nav metadata generation from directory structure.
I kind of wish there was just a plugin that did this naturally,
but, oh well.  I got it done in a relatively small number of lines.

The one concern I have here is it seems like if I did want to override
these properties again on a per-file basis, I'd... not be able to?
The computed properties are the dominant ones.

However, the only property where I think I care about that is title,
and as we can see in this diff already, I've just defined functions
that take this into account.

That approach should work fine for anything else that needing
customization, too: we can always add new properties that we'll alter
the computed property function to take account of.

(I wonder if we could override the eleventyComputed property itself
per directory or per file?  That might also work.  Haven't tried it.)

There's a new condition in the template to avoid adding empty links.
This is mostly to account for the home page, which otherwise
started showing up due to the broadness of application of this logic.
But it could of course be applied to exclude other things completely
from the nav tree as well.