Make breadcrumb links back up to parent pages also easily available.
Possibly redundant with the sidebar nav (once it has highlighting CSS,
anyway, which I haven't gotten around to yet).  But I think potentially
still also useful on its own.

This just directly mirrors the path structure.  I consider this sane.

The overflow may need some work for really long paths or really small screens.
Letting it expand might be trouble, because this layout requires some
coordinates to line up in order to have the nav area draw properly.
That... can be a problem for later.
I might actually consider aligning it to the right and cutting off the left
to just be perfectly valid in this case.