Content regarding libraries.
Just getting started.  There are not really any priors (or at least,
certainly not up-to-date priors) for this in the other docs and specs
repos, so I'm just starting from scratch even before merging those.

(Also, I just want some more content to fill out the nav menu,
as I continue to poke at the styling for that.)