Wip checkpoint on nav folding.
I'd like to have these menus be collapsible;
and I'd like to do that with CSS;
and this is definitely possible but it's a little fiddly.

Right now the styling is garish debug styling,
but functionally, fairly correct.

The main remaining incorrectness here (aside from the debug styling)
is that the checkbox is covering up the link,
making the link for those top level pages unclickable.

I've attempted some z-index shenanigans to get where I wanted,
but they don't appear to work, and I'm currently at a loss for why.

I may merge-ignore this (keep it in history, but drop it),
and revisit the topic later when we also have more sheer content
which will help make it easier to see how we need the nav to scale.