Several link fixes, stubs for more data model docs needed, new schema intro page.
Still more link fixes needed.  That ADL page was a doozy.

The data model pages also need filling in... but I think we may have
some good hunks in the specs repo already, so maybe I should get that
merged into the "_legacy" dir first and chomp it up.

Stopped myself on the new schema intro page for the same reason.
Pretty sure there's a lot of coverage of this already, it's just in
the specs repo, so it's not here to chomp up yet.