Reintroduce open-research page related to ADLs; finishes fixing all links from that document.  (Site should finally build again.)
Somewhat annoyed to discover this file disappeared back in commit
7068a17ee5dc4ed533afd9103054deb628936c6e in the docs repo...
which had a commit message of "fix: link to spec site for schemas".

Ironically, that commit broke a lot of things.  It's both a great case
to study for why "the link checker is important" -- many links were
broken -- and why "having multiple repos for this is a disaster" --
both content, and its history, was lost during that move.
The large churning of diff made it even easier to lose things, and
should have been avoidable.