Quickly outline the existence of some specs directories.
These will roughly parallel the docs directories of the same names.
Human-friendly and explanatory content goes in docs;
more mechanical sharp detail-oriented content goes in specs.

Haven't decided yet if the "known" directory will be useful for these.
For docs, it makes sense, because much of the content is general.
For the specs, it's unclear if we'll have "meta-specs"... or, if those
are things that we slide off into either the docs or the design dirs.
Probably I think we mostly slide that off into the design dirs,
and so in that case the "known" directory step might be less useful.