Porting first hunks of Data Model docs.
Though this content is coming from the specs repo... it's not super formal,
and so I'm putting it in the docs directories.

In general I would rather have the less formal information routed as docs,
and consider formal specifications a "todo".  I prefer this because it
will be easier to create a spec that starts with formal language than it
is to be stuck trying to evolve an informal doc into a formal one.
And we need the informal docs anyway!  So.  In the docs dir they go.