Port pathing docs.
Such as they are.

These are actually really limited in the quality of specification right now.

I'm intentionally leaving off some claims in the previous files about
this being well specified.  It is not.  Escaping mechanisms have not yet
been well reviewed or ratified.  Implementations are certainly not currently
unified around this.  Copying URI RFCs wholesale might well be the right thing
to do, but saying that it's a settled deal is just not an accurate claim about
the current practical state of affairs.