Introduce reusable child page listing system.
I want to deploy this on every page that has children.

Partly so they're not empty.

Partly so that the site is still navigable even if you don't ever open the nav menu.

Added one synopsys to prove it works and show how it can be done,
but I probably won't be able to write as many of these as are needed in one pass,
so added a placeholder text for the missing ones too.
Filling all these in might be something that's delegatable, even.