Clean up ADL page to at least use well-formatted todos.
Some content cross-linking is possible now too, so that improved a bit.
(This is one of the pages that came from the "docs" repo, and so
at the time it was written, it couldn't easily link to any of the content
that was at the time in the "specs" repo.  Sigh.  So much wasted time.
Anyway!  It's better now!)