Build a short(!) hello-world intro doc.
Move the what-is-ipld graphic here.

I think we may need to revist that graphic again.  It's good, but...
it's not perfect, either.  I'm concerned that it misleads people into
thinking that you need *either* schemas *or* ADLs in order to handle
Data Model data... and neither of those things are requirements.
The "plain" Data Model mode is perfectly useful.
This is something we debated previously while first crafting this
graphic, too, and realized it would probably be an issue, but couldn't
come up with an obvious solution.  A few moons later, I would say...
We're right that it's an issue; but I think we underestimated it,
in terms of how often is misleads people (it seems to be fairly often);
and so it's probably worth taking another think at how this particular
piece of information could be presented.