Migrate all the old design content to notebook.
Mostly, this involves taking some redundant directories out of the path.

At least one readme that was purely redundant also disappeared.

(I hate to say "I told you so" on the above, but... no, I don't;
I absolutely did, back at the time we created those directories,
state my bets that they were unnecessary and would never grow siblings.
Indeed: they did not.)

I've also taken the liberty of moving the "composites" doc to the
exploration report directory -- because that was what it was --
rather than continue to give it a weirdly lonely place above it all.
The date prefix assigned to it is from the git history of its
origin (and last touch; they're approximately the same).