Reorg schema docs: group into intro|features|using.
Lots of link fixing.

Content could probably be resettled a bit to match this,
and be improved by the journey, but that's a lot more work.
I hope just getting these groupings helps move forward a bit.

The text in the links and the actual link path are getting
more and more aligned by this migration, which is a good indicator.
(I'm not even making a particular point of that; it's just happening
fairly naturally.)

Slightly rewrote some intro material.  It's getting terser,
as it's able to move more into the subchapters, and even into
their synopsi.  (Synopsyses?  Oh my, what a word.)

Progress.