Quick port of the Quick Primer.
The origin of this doc is in https://hackmd.io/LHTTmGSWSvem4Wz2h_a39g ,
where it's already gotten some pretty good beta testing.

There exists some feedback on that upstream which is not yet fully
integrated, so we'll probably see some more changes to the file here
in the future, but for now let's at least get it tracked.

(I think some visitation on the visual presentation will also be
necessary... this document is very heavy on bullet points, and I don't
think the current CSS is making them ideally spaced for this content.)