Codec docs: "quick compare" entries, a page for dag-json, and a synopsys for everyone.
The 'quick compare' entries could probably be, well, quicker...
but I'm not quite willing to crush it all down into just thumbs up and
thumbs down emojis.

Yet.

Maybe in a future content pass we'll decide to compress this further.
There's certainly trends and repeated parts of the content in the
quick compare text already.