Prep for repo unification: move all content to a dir.
I will then merge this into the new site,
and slowly migrate content out of this dir piecemeal
(making sure all links in new content are coherent as I go).

(Okay, in this case, Oli will.  And already did, in fact.
I'm just going to rebase his work on top of this, for neato diffing.)