Crosslink schema docs and schema specs, and actually link schema specs content.
It's necessary to do the schema specs content linking manually,
because the 11ty autogen stuff we've been using for other kinds of
content only works for the markdown content pages (or stuff with
frontmatter, perhaps, but that still excludes these files).

Spartan, but makes the content reachable.  Previously it was not.