schema-schema: fix TypeUnion to have a 'members' field.
The rule is: the type definition must still be completely clear about
what will happen at the type level, even if the representation field
was stripped out or reset to a blank.  For union types: that means
the members list must be clear, and we cannot rely on the
representation strategy maps to contain it at some deeper level.

Yes, this means we must check for logical consistency between the
representation strategy specification and the members list.
But we have exactly the same kind of logical matchup work to do
for e.g. field names in representations for structs, etc,
so it is not unreasonable nor surprising to have that here for
union types too.

(There's also a logical mistake here that I'm propagating rather
than fixing: these shouldn't be 'TypeName', as Rod found and pointed
out and began to fix in https://github.com/ipld/ipld/pull/104 .
I'm going to circle back to addressing that in a future separate
commit, for clarity of diff and reviewability.)