schema-schema: TypeEnum.members is now stored as a list.
This is consistent with how TypeUnion.members is done.

It also removes the somewhat odd usage of "Null" in the schema-schema.
Although that recently actually became valid (because the Prelude was
finally actually explicitly documented, and includes the clause
`type Null unit representation null`), it was still sort of notablely
out of place because that was the _only_ place in the schema-schema
where a type name from the Prelude was referenced.
I'm happier with it gone.