schema-schema: introduce a struct in union representation declarations for stringprefix and bytesprefix.
This reserves some wiggle room for schema evolvablity:
it gives us room to add more fields in here later, if we choose to;
and we could do so without needing a breaking migration.
(With bare maps here, if we wanted to add "fields", we'd have no
way to distinguish the new fields from old map entries, and so
we'd have to introduce a breaking change somewhere *else* just to
give signalling space, and then do a migration, etc, etc.  Avoidable!)

---

We are *not* using this evolvablility today; just reserving
the room for it.

However, here's something we did consider adding, which provoked
this consideration, and just narrowly decided against:
an explicit joiner string for stringprefix unions.

Why?  Because one could imagine these two schemas:

```
type Foo union {
	| Bar "bar"
	| Baz "baz"
} representation stringprefix (":")
```

could also be written:

```
type Foo union {
	| Bar "bar:"
	| Baz "baz:"
} representation stringprefix
```

... and they would match the same stuff.

(And also, some early golang code actually did the former, even.)

A couple different influences tug in different directions here:

- an explicit joiner is not technically necessary --
  so, if it's not necessary, "clearly" we shouldn't have it, right?
- sometimes one doesn't *want* a uniform joiner --
  one can make a union with prefixes "$" and "/" and it should work.
- ... but, on the other hand: being able to split in a particular spot
  does make it easier for a validator to make clear error messages.

Ultimately, we're deciding the "it's not necessary, don't have it"
logic is the dominant one for now.  But we also feel that adding
the struct indirections here for evolvability wiggle room are still
worth it because there's very little cost to doing so.