schema-schema: switch to use of keyed unions.
This is a big switch.  At least, syntactically, in the DMT form.

It's a tiny switch, almost nothing at all, in the logical form.
(Neat how that works, isn't it?)

The primary driver for this is that it makes the schema-schema
use simpler features of the schema system.
This is something we're taking as explicit design goal for the
schema-schema from now on (and applying as a reason for some
significant "fixes" to previous choices like this one, too):

**The schema-schema should intentionally use only as many of
the features of schemas as necessary**, and **it should prefer using
the features that we feel are reasonable to expect to be implemented
early in a new IPLD library in a new language**.

That design rule means inline unions are _out_.

(I actually added documentation about that design goal a few commits
ago, in prose, in one of the other specs pages;
now it's being enacted!)

(A quick rant about inline unions:

Inline unions are one of the most complicated things in the whole
system to write a correct parser for.  They're *riddled* with
edge cases and tricky corners.  And making them efficient?  Fwoosh.
Efficiently parsing an inline union is a absolute nightmare;
there's possibilities for lookahead and lookback and reprocess galore.
They're orders of magnitude more complicated than any other feature.
Frankly, I'd never have specified them at all if in a total greenfield;
the reason the inline union strategy is present in IPLD Schemas at all
is because of our broader goal to "describe preexiting data,
and be able to embrace conventions which were already popular".
Outside of compatibility missions, I would now never recommend the
use of inline union strategies in new protocols.

And on top of that -- personally -- I think inline unions suck to read.
Keyed unions have become the most readily legible thing to me,
after a few months of reading both.  In my opinion, after this diff,
the schema-schema JSON document is significantly less bewildering.

... Right.  Rant complete.  Moving on.)

The diff here is unfortunately large (nearly total, in the JSON),
because keyed unions add a level of indentation where inline did not.
Still: worth it.