schema-schema: TypeLink.expectedType is TypeName, not string; also, docs.
That expectedType should be TypeName is clear enough.

I took the opportunity to revisit the doc coments on TypeLink
significantly.  We can talk much more clearly about the DSL syntax now!
That's exciting.  (We were frankly winging it the last time we updated
these docs, I think.)  Now, it's quite clear: the 'Any' type is
a well-specified thing (named in the Prelude); its behavior is clearly
the right thing for our default here, without any additional magic
(because the 'any' typekind specification now does the right thing);
and the `Link` syntax is also still defined, but now without magic,
because it's again in the Prelude (it's just `type Link &Any`).
Everything in the DSL we ever wanted now
A) works, and B) without magic nor special cases.  *Sweet.*