schema-schema: various docs improvements.
I updated some of the prose, mostly around representation strategies,
to how we'd write it today.  Much of our terminology has gotten
clearer as we've polished it over time.

This is not a complete pass over the whole schema-schema;
just some improvements of opportunity I noticed while in the area.