schema-schema: rename Type->TypeDefn, and similar.
More specifically:

1. `Type` is now `TypeDefn`.
2. All `Type{typekind}` are now `TypeDefn{typekind}`.
3. And also, yes, `TypeCopy`, that odd duckling, is now `TypeDefnCopy`.

To explain why this new terminology makes sense,
here are two pieces of prose.
The is first written with our old terminology,
and the second with the new terms:

> "copy" is not present in the TypeKind enum, while it _is_ present in
> the Type union.  That's not a typo, and kind of accurately indicates
> the subtlety: you can define a type using TypeCopy, but it's
> not actually a typekind.

... and now read that paragraph again with the new names:

> "copy" is not present in the TypeKind enum, while it _is_ present in
> the TypeDefn union.  That's not a typo, and kind of accurately indicates
> the subtlety: you can define a type using TypeDefnCopy, but it's
> not actually a typekind.

... clearer, right?

These renames are a decent diff size in the schema-schema... but,
due to the joys of structural typing, that's it: these changes
have _no effect_ on the actual DMT of schemas.  It's _just_ an
internal change to naming.  So, the `examples.ipldsch.json` file?
The diff indeed doesn't touch that file, and that's not a bug.