schema-schema: drop SchemaMap type (it's easy to make anonymous); and move Schema to top of document; and fix docs.
This makes the top of the JSON document... somewhat gorgeous:

```
{
  "types": {
    "Schema": {
      "struct": {
        "fields": {
          "types": {
            "type": {
              "map": {
```

Aren't self-describing schemas Fun?  :D

(It's always been this heady, but it just got slightly more visually so,
as switching an anonymous map for the type definitions makes it deeper,
and shifting the root type of the schema (which is, of course, called "Schema")
to the top of the document makes this... wonderful cascade effect come out.
Whee :))