schema-schema: fix quoting typos around implicits in the DSL form.
Which kind of value we have there (aka which member of AnyScalar a
parser should yield) should be obvious from the syntax.

So, quotation marks should be present for strings; for bool literals,
there should be no quotations; etc.

(If the member of AnyScalar that a parser finds doesn't actually match
the kind for the type for the field that the implicit clause is
describing, that's a logical consistency check that that should be
noticed during a "compile" phase -- not necessarily something that
the DSL parser should need to worry about.)