specs/schemas: regenerate schema-schema.ipldsch.json
This uses an updated version of schema/dsl and schema/dmt,
which I haven't posted as a pull request yet.

The new ipldsch.json file is the result of parsing the ipldsch file
and encoding it with IPLD's JSON codec.
The JSON codec keeps the original key order and indents.

The following changes show up:

1) Two "inline" union representations were changed in the ipldsch file,
   but we had not updated the json accordingly.

2) Schema's "advanced" field is optional.

3) ListRepresentation_List is no longer declared nor used.

4) The "string" enum representations now map from string to enum value,
   which closely matches the other representation maps.