specs/schemas: port several fixtures to testmark format, along with updating them.
I've added a lot of prose explainer interspersed between data hunks.

I chose to combine a bunch of the basic things (especially, scalar kinds like int and float)
into one file.  This might be stylistic (and I could be convinced to split it back out),
but made sense to me, and felt relatively reasonable-looking in testmark (because
it's fairly normative to start hunk names in testmark with a test name prefix and
have several fixtures in the same file that way), so I gave it a spin.

I've ditched the actual-vs-expected distinction in example blocks.
In almost all cases, they were equal.  The only places where they differed
where where there were comments speculating if various coersions like string-to-int
or float-to-int were supposed to happen.  In all cases, the answer to those
questions is "no".  So that simplifies things nicely.
(Okay, there's one more exception around json floats, but... I let this be lost;
I think that's testing json codec normalization, not schemas; it doesn't belong here.)

I renamed "blocks" to "match" and "badBlocks" to "nomatch".  We could bikeshed
these to something else (I'm not super attached to my picks), or revert to those
previous terms; I just felt a little strange at the word "block" floating around.

I'm trying to avoid tackling this too much, at the moment but it may be worth
a brief mention that some of the fixtures for data that matches feel like they
might be probing tangents.  For example, a bunch of the float fixtures seem
like they're sanity checking the json float parser as much as they're saying
anything about the schema matching. I'm not sure these are optimally useful.
But they also don't hurt, so I've (mostly) kept them.

I made some target-of-opportunity fixes, such as replacing fixtures about null with
updated fixtures about unit (which follows https://github.com/ipld/ipld/pull/136 ).

I put all this in a dir called "fixtures" because that's what the similar dir is
called in the selector specs, and in most of the codec specs that have them.
(We're not 100% consistent on this either, in fairness.  For the CAR specs,
the dir is called "fixture" -- no "s".  Ah well.  Aiming towards the center of gravity.)

So far this diff just covers a few of the files; I've removed the content
I've ported, but there's a lot to go.